Refactor ViewModel usage in navigation
The `SpeciesDetailViewModel` is now directly retrieved using `koinViewModel()` in the `DetailSpecies` screen composable.

In the `DetailImages` screen composable, a try-catch block has been added when retrieving the `SpeciesDetailViewModel` to handle potential `IllegalArgumentException` if the back stack entry for `Screen.DetailSpecies.route` is not found, falling back to the current `backStackEntry` in such cases. A TODO comment has also been added to consider a separate ViewModel for image details.