Add description sections to Family and Species
This commit introduces a new `DescriptionSection` data class and adds a list of these sections to the `Description` model.

The `FamilyDetailMapper` and `SpeciesDetailMapper` have been updated to populate these new description sections from the corresponding fields in the API response. This allows for a more structured representation of descriptive information for families and species.